feat(ios): Call `onOpenWindow` for links with `target="_blank"`
On iOS, default behavior is to navigate the WebView to the target URL
when a link with `target="_blank"` is clicked

In some scenario we want to be able to intercept this from react-native
and to handle the link by ourselves (i.e. to open it into an InApp
Browser)

The `onOpenWindow` event is now called in this scenario. Then the
default "open browser" scenario does not occur

Note that previous implementation only worked in apparence, but the
WebView ref url would still be updated with new URL even if
`return nil` is done in `createWebViewWithConfiguration`. So we want to
intercept this earlier in `decidePolicyForNavigationAction` when the
WebView did not update its ref yet